refactor: move checkout to node-app
