+++ imports/node-app/plugins/checkout/i18n/de.json
