+++ imports/node-app/plugins/checkout/i18n/es.json
