+++ imports/node-app/plugins/checkout/i18n/fr.json
