+++ imports/node-app/plugins/checkout/i18n/it.json
