+++ imports/node-app/plugins/checkout/i18n/nb.json
